+++ app.py
@@ -9,8 +9,9 @@ from raven.contrib.flask import Sentry
 from sqlalchemy.exc import OperationalError, TimeoutError
 
 from config import DevConfig, ProdConfig
-from init_db import db, ma
+from init_db import db
 from utils.xlsx_reader import get_books, get_magazines
+from utils.create_admin_user import create_super_user
 from views.book import library_books
 from views.book_borrowing_dashboard import library_book_borrowing_dashboard
 from views.index import library
@@ -19,10 +20,10 @@ mail = Mail()
 sentry = Sentry()
 client = Client()
 
-if os.getenv('APP_SETTINGS', '') == 'prod':
-    config_env = ProdConfig
-else:
+if os.getenv('FLASK_ENV', '') == 'development':
     config_env = DevConfig
+else:
+    config_env = ProdConfig
 
 
 def create_app(config=config_env):
@@ -34,7 +35,6 @@ def create_app(config=config_env):
     app.secret_key = os.urandom(24)
     mail.init_app(app)
     db.init_app(app)
-    ma.init_app(app)
     wait_for_db(app)
     return app
 
@@ -61,10 +61,18 @@ app = create_app()
 migrate = Migrate(app, db)
 
 
-@app.cli.command(with_appcontext=True)
+@app.cli.command('load_xls_into_db', with_appcontext=True)
 def load_xls_into_db():
     get_magazines('./library_example.xlsx')
     get_books('./library_example.xlsx')
 
 
 app.cli.add_command(load_xls_into_db)
+
+
+@app.cli.command('create_admin', with_appcontext=True)
+def create_admin():
+    create_super_user()
+
+
+app.cli.add_command(create_admin)