+++ forms/book.py
@@ -34,37 +34,34 @@ class LibraryItemForm(FlaskForm):
                                       render_kw=({'class': 'inputs',
                                                   'id': 'table_of_contest',
                                                   'placeholder':
-                                                      'Table of contents',
-                                                  'disabled': True}))
+                                                      'Table of contents'}))
 
     language = SelectField('Language',
                            choices=languages,
                            validators=[check_language],
                            render_kw=({
-                               'class': 'custom-select mb-2 mr-sm-2 mb-sm-0',
+                               'class': 'inputs custom-select '
+                                        'mb-2 mr-sm-2 mb-sm-0',
                                'id': 'mySelect',
-                               'placeholder': 'Language',
-                               'disabled': True}))
+                               'placeholder': 'Language'}))
 
     category = SelectField('Category',
                            choices=categories,
                            validators=[check_category],
                            render_kw=({
-                               'class': 'custom-select mb-2 mr-sm-2 mb-sm-0',
+                               'class': 'inputs custom-select'
+                                        ' mb-2 mr-sm-2 mb-sm-0',
                                'id': 'mySelect',
-                               'placeholder': 'Category',
-                               'disabled': True}))
+                               'placeholder': 'Category'}))
 
     tag = StringField('Tag',
                       render_kw=({'class': 'inputs',
                                   'id': 'tag',
-                                  'placeholder': 'Tag',
-                                  'disabled': True}))
+                                  'placeholder': 'Tag'}))
     description = TextAreaField('Description',
                                 render_kw=({'class': 'inputs',
                                             'id': 'description',
-                                            'placeholder': 'Description',
-                                            'disabled': True}))
+                                            'placeholder': 'Description'}))
 
     pub_date = SelectField('Year of publication',
                            choices=[(str(year), str(year))
@@ -73,15 +70,14 @@ class LibraryItemForm(FlaskForm):
                                           datetime.now().year + 1)],
                            validators=[check_pub_date],
                            render_kw=({
-                               'class': 'custom-select mb-2 mr-sm-2 mb-sm-0',
+                               'class': 'inputs custom-select '
+                                        'mb-2 mr-sm-2 mb-sm-0',
                                'id': 'mySelect',
-                               'placeholder': 'Year of publication',
-                               'disabled': True}))
+                               'placeholder': 'Year of publication'}))
 
     submit = SubmitField('Update',
                          render_kw=({'class': 'btn btn-primary submits',
-                                     'id': 'button',
-                                     'disabled': True}))
+                                     'id': 'button'}))
 
 
 class BookForm(LibraryItemForm):
@@ -91,61 +87,51 @@ class BookForm(LibraryItemForm):
                                     title_book_exists],
                         render_kw=({'class': 'inputs',
                                     'id': 'title',
-                                    'placeholder': 'Title',
-                                    'disabled': True}))
+                                    'placeholder': 'Title'}))
 
     isbn = StringField('ISBN number',
                        validators=[DataRequired(), check_isbn],
                        render_kw=({'class': 'inputs',
                                    'id': 'isbn',
-                                   'placeholder': 'ISBN number',
-                                   'disabled': True}))
+                                   'placeholder': 'ISBN number'}))
 
     original_title = StringField('Original title',
                                  render_kw=({'class': 'inputs',
                                              'id': 'original_title',
-                                             'placeholder': 'Original title',
-                                             'disabled': True}))
+                                             'placeholder': 'Original title'}))
     publisher = StringField('Publisher',
                             validators=[DataRequired()],
                             render_kw=({'class': 'inputs',
                                         'id': 'publisher',
-                                        'placeholder': 'Publisher',
-                                        'disabled': True}))
+                                        'placeholder': 'Publisher'}))
 
     # Authors
     first_name = StringField('First name',
                              validators=[DataRequired(), check_author],
                              render_kw=({'class': 'inputs',
-                                         'placeholder': 'First Name',
-                                         'disabled': True}))
+                                         'placeholder': 'First Name'}))
     surname = StringField('Surname',
                           validators=[DataRequired(), check_author],
                           render_kw=({'class': 'inputs',
-                                      'placeholder': 'Surname',
-                                      'disabled': True}))
+                                      'placeholder': 'Surname'}))
 
     first_name_1 = StringField('First_name_1',
                                validators=[check_author],
                                render_kw=({'class': 'inputs',
-                                           'placeholder': 'First Name 1',
-                                           'disabled': True}))
+                                           'placeholder': 'First Name 1'}))
     surname_1 = StringField('Surname_1',
                             validators=[check_author],
                             render_kw=({'class': 'inputs',
-                                        'placeholder': 'Surname 1',
-                                        'disabled': True}))
+                                        'placeholder': 'Surname 1'}))
 
     first_name_2 = StringField('First name 2',
                                validators=[check_author],
                                render_kw=({'class': 'inputs',
-                                           'placeholder': 'First Name 2',
-                                           'disabled': True}))
+                                           'placeholder': 'First Name 2'}))
     surname_2 = StringField('Surname 2',
                             validators=[check_author],
                             render_kw=({'class': 'inputs',
-                                        'placeholder': 'Surname 2',
-                                        'disabled': True}))
+                                        'placeholder': 'Surname 2'}))
 
 
 class MagazineForm(LibraryItemForm):
@@ -155,18 +141,21 @@ class MagazineForm(LibraryItemForm):
                                     render_kw=({'class': 'inputs',
                                                 'id': 'title_of_magazine',
                                                 'placeholder':
-                                                    'Title of magazine',
-                                                'disabled': True}))
+                                                    'Title of magazine'}))
 
     issue = StringField('Issue',
                         render_kw=({'class': 'inputs',
                                     'id': 'issue',
-                                    'placeholder': 'Issue',
-                                    'disabled': True}))
+                                    'placeholder': 'Issue'}))
 
 
-class MixedForm(MagazineForm, BookForm):
-    submit = SubmitField('Create',
-                         render_kw=({'class': 'btn btn-primary submits',
-                                     'id': 'button',
-                                     'disabled': True}))
+class AddNewItemBookForm(BookForm):
+    submit1 = SubmitField('Create Book',
+                          render_kw=({'class': 'btn btn-primary submits',
+                                     'id': 'button1'}))
+
+
+class AddNewItemMagazineForm(MagazineForm):
+    submit2 = SubmitField('Create Magazine',
+                          render_kw=({'class': 'btn btn-primary submits',
+                                     'id': 'button2'}))