+++ serializers/wishlist.py
@@ -1,7 +1,0 @@
-from init_db import ma
-from models.wishlist import WishListItem
-
-
-class WishListItemSchema(ma.ModelSchema):
-    class Meta:
-        model = WishListItem