+++ tests/test_book.py
@@ -1,6 +1,8 @@
-from flask import url_for
-from datetime import date
+from flask import session
+from datetime import date, datetime
 from random import choice, randint
+
+from flask import url_for
 from mimesis import Generic
 import pytest
 
@@ -8,22 +10,20 @@ from models import Author, Book, Tag, Magazine, LibraryItem
 from forms.book import BookForm, MagazineForm
 
 
-def test_add_book(view_book, client):
+def test_add_book(view_book, client, login_form_admin_credentials):
     view_book.radio.data = 'book'
-
+    client.post(url_for('library.login'),
+                data=login_form_admin_credentials.data)
     client.post(url_for('library_books.add_book'),
                 data=view_book.data,
                 follow_redirects=True)
-
-    del view_book.issue
-    del view_book.title_of_magazine
 
     author = Author.query.filter_by(first_name=view_book.first_name.data,
                                     last_name=view_book.surname.data).first()
     if not author:
         assert False, "Data validation failed"
-    assert view_book.first_name.data == author.first_name and view_book. \
-        surname.data == author.last_name, \
+    assert view_book.first_name.data == author.first_name \
+        and view_book.surname.data == author.last_name, \
         "First and last name of the first author is not" \
         " the same as given at the entrance"
 
@@ -58,61 +58,63 @@ def test_add_book(view_book, client):
 
     assert tag.name in view_book.tag.data, \
         "Tags ane not the same"
+    session.clear()
 
 
-def test_add_magazine(view_book, client):
-    view_book.radio.data = 'magazine'
-
+def test_add_magazine(view_magazine, client, login_form_admin_credentials):
+    view_magazine.radio.data = 'magazine'
+    client.post(url_for('library.login'),
+                data=login_form_admin_credentials.data)
     client.post(url_for('library_books.add_book'),
-                data=view_book.data,
+                data=view_magazine.data,
                 follow_redirects=True)
 
-    del view_book.publisher
-    del view_book.original_title
-    del view_book.isbn
-    del view_book.title
-
-    magazine = Magazine.query.filter_by(title=view_book.title_of_magazine.data,
-                                        issue=view_book.issue.data[0]).first()
+    magazine = Magazine.query\
+        .filter_by(title=view_magazine.title_of_magazine.data,
+                   issue=view_magazine.issue.data[0]).first()
     if not magazine:
         assert False, "Data validation failed"
-    assert view_book.title_of_magazine.data == magazine.title, \
+    assert view_magazine.title_of_magazine.data == magazine.title, \
         "The title of the book is not the same as given at the entrance "
-    assert view_book.table_of_contents.data == magazine.table_of_contents, \
+    assert \
+        view_magazine.table_of_contents.data == magazine.table_of_contents, \
         "The table of content is not the same as given at the entrance "
-    assert view_book.language.data == magazine.language, \
+    assert view_magazine.language.data == magazine.language, \
         "Language is not the same as given at the entrance "
-    assert view_book.category.data == magazine.category, \
+    assert view_magazine.category.data == magazine.category, \
         "Category of book is not the same as given at the entrance "
-    assert view_book.description.data == magazine.description, \
+    assert view_magazine.description.data == magazine.description, \
         "The book description is not the same as given at the entrance "
-    assert date(year=int(view_book.pub_date.data),
+    assert date(year=int(view_magazine.pub_date.data),
                 month=1,
                 day=1) == magazine.year, \
         "The year of publication is not the same as given at the entrance "
-    assert view_book.issue.data[0] == magazine.issue, \
+    assert view_magazine.issue.data[0] == magazine.issue, \
         "The book description is not the same as given at the entrance "
 
-    tag = Tag.query.filter_by(name=view_book.tag.data[0]).first()
+    tag = Tag.query.filter_by(name=view_magazine.tag.data[0]).first()
     if not tag:
         assert False, "Data validation failed"
-    assert tag.name in view_book.tag.data, \
+    assert tag.name in view_magazine.tag.data, \
         "Tags ane not the same"
+    session.clear()
 
 
-def test_add_the_same_book(view_book, client):
+def test_add_the_same_book(view_magazine, client):
     client.post(url_for('library_books.add_book'),
-                data=view_book.data,
+                data=view_magazine.data,
                 follow_redirects=True)
 
     client.post(url_for('library_books.add_book'),
-                data=view_book.data,
+                data=view_magazine.data,
                 follow_redirects=True)
-    assert not bool(view_book.errors), \
+    assert not bool(view_magazine.errors), \
         "Two same books have been added."
 
 
 """Testing separated validators"""
+
+
 @pytest.mark.parametrize("values, result", [
     ("", True),
     (".", False),
@@ -124,6 +126,7 @@ def test_add_the_same_book(view_book, client):
     ("A.Adsa", True),
     ("A.AdsaA", False),
     ("Paweł", True)
+
 ])
 def test_check_author(view_book, values, result):
     view_book.first_name_1.data = values
@@ -188,8 +191,8 @@ def test_check_isbn(view_book, values, result):
 @pytest.mark.parametrize("values, result", [
     ("1969", False),
     ("1970", True),
-    ("2018", True),
-    ("2019", False),
+    (str(datetime.now().year), True),
+    (str(datetime.now().year + 1), False),
     (2005, False)
 
 ])
@@ -204,7 +207,7 @@ def test_check_pub_date(view_book, values, result):
 # End of Testing separated validators
 
 
-def test_update_book(view_edit_book, client):
+def test_update_book(view_edit_book, client, login_form_admin_credentials):
     languages = ['polish', 'english', 'other']
     categories = ['developers', 'managers',
                   'magazines', 'other']
@@ -227,7 +230,8 @@ def test_update_book(view_edit_book, client):
         publisher=g.business.company(),
         pub_date=str(randint(1970, 2018))
     )
-
+    client.post(url_for('library.login'),
+                data=login_form_admin_credentials.data)
     client.post(url_for('library_books.edit_book', item_id=item.id),
                 data=form.data,
                 follow_redirects=True)
@@ -259,9 +263,11 @@ def test_update_book(view_edit_book, client):
         "Book pub_date has not been updated"
     assert tmp_item.tags[0].name == form.tag.data[0], \
         "Book tags has not been updated"
+    session.clear()
 
 
-def test_update_magazine(view_edit_magazine, client):
+def test_update_magazine(view_edit_magazine, client,
+                         login_form_admin_credentials):
     languages = ['polish', 'english', 'other']
     categories = ['developers', 'managers',
                   'magazines', 'other']
@@ -280,7 +286,8 @@ def test_update_magazine(view_edit_magazine, client):
         description=g.text.sentence(),
         pub_date=str(randint(1970, 2018))
     )
-
+    client.post(url_for('library.login'),
+                data=login_form_admin_credentials.data)
     client.post(url_for('library_books.edit_book', item_id=item.id),
                 data=form.data,
                 follow_redirects=True)
@@ -302,3 +309,4 @@ def test_update_magazine(view_edit_magazine, client):
         "Book pub_date has not been updated"
     assert tmp_item.tags[0].name == form.tag.data[0], \
         "Book tags has not been updated"
+    session.clear()