+++ views/index.py
@@ -41,9 +41,13 @@ from models import LibraryItem
 from models.library import RentalLog, Copy, BookStatus
 from models.users import User
 from models.wishlist import WishListItem, Like
+from models.decorators_roles import (
+    require_role,
+    require_logged_in,
+    require_not_logged_in
+)
 from send_email import send_confirmation_email, send_password_reset_email
 from send_email.emails import send_email
-from serializers.wishlist import WishListItemSchema
 
 library = Blueprint('library', __name__,
                     template_folder='templates')
@@ -55,6 +59,7 @@ def index():
 
 
 @library.route('/login', methods=['GET', 'POST'])
+@require_not_logged_in()
 def login():
     if request.method == 'GET':
         if 'logged_in' in session:
@@ -71,7 +76,6 @@ def login():
     else:
         form = LoginForm()
         try:
-
             if form.validate_on_submit():
                 data = User.query.filter_by(email=form.email.data).first()
                 if (data is not None and
@@ -106,6 +110,7 @@ def login():
 
 
 @library.route('/registration', methods=['GET', 'POST'])
+@require_not_logged_in()
 def registration():
     if request.method == 'GET':
         form = RegistrationForm()
@@ -122,7 +127,6 @@ def registration():
                     return render_template('message.html',
                                            message_title=message_title,
                                            message_body=message_body)
-
                 else:
                     new_user = User(
                         email=form.email.data,
@@ -151,8 +155,8 @@ def registration():
 
 
 @library.route('/search', methods=['GET'])
+@require_logged_in()
 def search():
-
     try:
         user = User.query.get(session['id'])
         admin = user.has_role('ADMIN')
@@ -160,25 +164,24 @@ def search():
         abort(401)
     except Exception:
         abort(500)
-
     if request.method == 'GET':
         if not request.args or not request.args.get('query'):
             form = SearchForm()
             page = request.args.get('page', 1, type=int)
             try:
                 paginate_query = LibraryItem.query.order_by(
-                    LibraryItem.title.asc()).paginate(page, 10, False)
-
+                    LibraryItem.title.asc()).paginate(page,
+                                                      error_out=True,
+                                                      max_per_page=10)
                 output = [d.serialize() for d in paginate_query.items]
                 return render_template('search.html',
                                        all_query=output,
                                        admin=admin,
                                        pagination=paginate_query,
                                        endpoint='library.search',
-                                       form=form,)
+                                       form=form)
             except RuntimeError:
                 return ErrorMessage.message('Cannot connect to database!')
-
         elif request.args.get('query'):
             form = SearchForm()
             query_str = request.args.get('query')
@@ -186,22 +189,20 @@ def search():
             try:
                 paginate_query = (
                     LibraryItem.query.filter(LibraryItem.title.ilike(
-                        '%{}%'.format(query_str)))).paginate(page, 10, False)
-
+                        '%{}%'.format(query_str)))).paginate(page,
+                                                             error_out=True,
+                                                             max_per_page=10)
                 output = [d.serialize() for d in paginate_query.items]
-
             except RuntimeError:
                 return ErrorMessage.message('Cannot connect to database!')
-
             return render_template('search.html',
                                    all_query=output,
                                    pagination=paginate_query,
                                    endpoint='library.search',
                                    admin=admin,
-                                   form=form,)
-
+                                   form=form, )
         else:
-            abort(500)
+            abort(405)
 
 
 @library.route('/contact', methods=['GET', 'POST'])
@@ -209,7 +210,7 @@ def contact():
     form = ContactForm()
     if form.validate_on_submit():
         email_template = open(
-            './templates/contact_confirmation.html', 'r').read()
+            './templates/email_template/contact_confirmation.html', 'r').read()
         try:
             send_email(
                 'Contact confirmation, title: ' + form.title.data,
@@ -217,17 +218,16 @@ def contact():
                 [form.email.data],
                 None,
                 email_template)
-
             send_email(
                 'Contact form: ' + form.title.data,
                 DevConfig.MAIL_USERNAME,
                 [DevConfig.MAIL_USERNAME],
                 'Send by: ' + form.email.data + '\n\n' + form.message.data,
                 None)
-            return SuccessMessage\
+            return SuccessMessage \
                 .message('Your email has been sent to administrator!')
         except TimeoutError:
-            return ErrorMessage\
+            return ErrorMessage \
                 .message('Oops, '
                          'some problem occurred'
                          ' and your email has not been sent ')
@@ -238,6 +238,7 @@ def contact():
 
 
 @library.route('/logout')
+@require_logged_in()
 def logout():
     session.clear()
     return render_template('index.html')
@@ -256,9 +257,7 @@ def confirm_email(token):
         return render_template('message.html',
                                message_title=message_title,
                                message_body=message_body)
-
     user = User.query.filter_by(email=email).first()
-
     if user.active:
         message_body = 'Account already confirmed. Please login.'
         message_title = 'Error!'
@@ -277,6 +276,7 @@ def confirm_email(token):
 
 
 @library.route('/reset', methods=['GET', 'POST'])
+@require_not_logged_in()
 def reset():
     form = ForgotPass()
     if form.validate_on_submit():
@@ -320,9 +320,7 @@ def reset_with_token(token):
         return render_template('message.html',
                                message_title=message_title,
                                message_body=message_body)
-
     form = PasswordForm()
-
     if form.validate_on_submit():
         try:
             user = User.query.filter_by(email=email).first_or_404()
@@ -332,13 +330,11 @@ def reset_with_token(token):
             return render_template('message.html',
                                    message_title=message_title,
                                    message_body=message_body)
-
         user.password_hash = generate_password_hash(form.password.data)
         db.session.add(user)
         db.session.commit()
         flash('Your password has been updated!', 'success')
         return redirect(url_for('library.login'))
-
     return render_template('reset_password_with_token.html',
                            form=form,
                            token=token,
@@ -346,31 +342,31 @@ def reset_with_token(token):
 
 
 @library.route('/reservation/<copy_id>')
+@require_logged_in()
 def reserve(copy_id):
-    if 'logged_in' in session:
-        try:
-            copy = Copy.query.get(copy_id)
-            copy.available_status = BookStatus.RESERVED
-            res = RentalLog(
-                copy_id=copy_id,
-                user_id=session['id'],
-                book_status=BookStatus.RESERVED,
-                reservation_begin=datetime.now(tz=pytz.utc),
-                reservation_end=datetime.now(
-                    tz=pytz.utc) + timedelta(minutes=2))
-            db.session.add(res)
-            db.session.commit()
-            flash('Pick up the book within two days!')
-        except IntegrityError:
-            abort(500)
+    try:
+        copy = Copy.query.get(copy_id)
+        copy.available_status = BookStatus.RESERVED
+        res = RentalLog(
+            copy_id=copy_id,
+            user_id=session['id'],
+            book_status=BookStatus.RESERVED,
+            reservation_begin=datetime.now(tz=pytz.utc),
+            reservation_end=datetime.now(
+                tz=pytz.utc) + timedelta(minutes=2))
+        db.session.add(res)
+        db.session.commit()
+        flash('Pick up the book within two days!')
+    except IntegrityError:
+        abort(500)
     return redirect(url_for(
         'library_book_borrowing_dashboard.book_borrowing_dashboad'))
 
 
 @library.route('/check_reservation_status_db')
 def check_reservation_status_db():
-    reserved_list = db.session.query(RentalLog)\
-        .filter(RentalLog.book_status == BookStatus.RESERVED)\
+    reserved_list = db.session.query(RentalLog) \
+        .filter(RentalLog.book_status == BookStatus.RESERVED) \
         .all()
     db.session.query(Copy).filter(
         Copy.id.in_([obj.copy_id for obj in reserved_list])
@@ -378,14 +374,15 @@ def check_reservation_status_db():
         {Copy.available_status: BookStatus.RETURNED},
         synchronize_session='fetch'
     )
-    db.session.query(RentalLog)\
-        .filter(RentalLog.book_status == BookStatus.RESERVED)\
+    db.session.query(RentalLog) \
+        .filter(RentalLog.book_status == BookStatus.RESERVED) \
         .update({RentalLog.book_status: BookStatus.RETURNED})
     db.session.commit()
     return "OK"
 
 
 @library.route('/remove_item/<int:item_id>', methods=['GET', 'POST'])
+@require_role('ADMIN')
 def remove_item(item_id):
     try:
         user = User.query.get(session['id'])
@@ -413,14 +410,10 @@ def remove_item(item_id):
 
 @library.route('/remove_copy/<int:item_id>/<int:copy_id>',
                methods=['GET', 'POST'])
+@require_role('ADMIN')
 def remove_copy(item_id, copy_id):
-    try:
-        user = User.query.get(session['id'])
-        admin = user.has_role('ADMIN')
-    except KeyError:
-        abort(401)
-    except Exception:
-        abort(500)
+    user = User.query.get(session['id'])
+    admin = user.has_role('ADMIN')
     form = RemoveForm()
     item = LibraryItem.query.get_or_404(item_id)
     copy = Copy.query.filter_by(id=copy_id).first_or_404()
@@ -442,22 +435,59 @@ def remove_copy(item_id, copy_id):
 
 
 @library.route('/wishlist', methods=['GET', 'POST'])
+@require_logged_in()
 def wishlist():
-    try:
-        user = User.query.get(session['id'])
-        admin = user.has_role('ADMIN')
-    except KeyError:
-        abort(401)
-    except Exception:
-        abort(500)
-
-    data = db.session.query(WishListItem).all()
-    wish_list_schema = WishListItemSchema(many=True)
-    output = wish_list_schema.dump(data)
-    return render_template('wishlist.html', wishes=output, admin=admin)
+    user = User.query.get(session['id'])
+    admin = user.has_role('ADMIN')
+    if request.method == 'GET':
+        if not request.args or not request.args.get('query'):
+            form = SearchForm()
+            page = request.args.get('page', 1, type=int)
+            if db.session.query(WishListItem).first() is None:
+                return render_template('wishlist.html', admin=admin)
+            try:
+                data = (
+                    WishListItem.query.order_by(
+                        WishListItem.likes_count.desc()).order_by(
+                        WishListItem.title.asc()).paginate(page,
+                                                           error_out=True,
+                                                           max_per_page=5))
+            except RuntimeError:
+                return ErrorMessage.message('Cannot connect to database!')
+            output = [d.serialize() for d in data.items]
+            return render_template('wishlist.html',
+                                   wishes=output,
+                                   admin=admin,
+                                   pagination=data,
+                                   endpoint='library.wishlist',
+                                   form=form, )
+        elif request.args.get('query'):
+            form = SearchForm()
+            query_str = request.args.get('query')
+            page = request.args.get('page', 1, type=int)
+            try:
+                data = (
+                    WishListItem.query.filter(WishListItem.title.ilike(
+                        '%{}%'.format(query_str))).order_by(
+                        WishListItem.likes_count.desc()).order_by(
+                        WishListItem.title.asc()).paginate(page,
+                                                           error_out=True,
+                                                           max_per_page=5))
+            except RuntimeError:
+                return ErrorMessage.message('Cannot connect to database!')
+            output = [d.serialize() for d in data.items]
+            return render_template('wishlist.html',
+                                   wishes=output,
+                                   admin=admin,
+                                   pagination=data,
+                                   endpoint='library.wishlist',
+                                   form=form, )
+    else:
+        abort(405)
 
 
 @library.route('/add_wish', methods=['GET', 'POST'])
+@require_logged_in()
 def add_wish():
     form = WishlistForm()
     if form.validate_on_submit():
@@ -468,7 +498,6 @@ def add_wish():
                                          pub_year=datetime.strptime(
                                              form.pub_date.data,
                                              "%Y").date())
-
             db.session.add(new_wish_item)
             db.session.commit()
             return redirect(url_for('library.wishlist'))
@@ -478,6 +507,7 @@ def add_wish():
 
 
 @library.route('/add_like', methods=['GET', 'POST'])
+@require_logged_in()
 def add_like():
     wish_id = request.form['wish_id']
     user = User.query.filter_by(id=session['id']).first()
@@ -492,20 +522,22 @@ def add_like():
         except exc.SQLAlchemyError:
             return ErrorMessage.message(error_body='Oops something went wrong')
     return json.dumps({'num_of_likes': len(WishListItem.query
-                                                       .filter_by(id=wish_id)
-                                                       .first().likes)})
+                                           .filter_by(id=wish_id)
+                                           .first().likes)})
 
 
 @library.route('/delete_wish/<int:wish_id>', methods=['GET', 'POST'])
+@require_role('ADMIN')
 def delete_wish(wish_id):
     try:
         WishListItem.delete_wish(wish_id)
     except exc.SQLAlchemyError:
-            return ErrorMessage.message(error_body='Oops something went wrong')
+        return ErrorMessage.message(error_body='Oops something went wrong')
     return redirect(url_for('library.wishlist'))
 
 
 @library.route('/item_description/<int:item_id>')
+@require_logged_in()
 def item_description(item_id):
     try:
         user = User.query.get(session['id'])
@@ -516,11 +548,9 @@ def item_description(item_id):
         abort(500)
     item = LibraryItem.query.get_or_404(item_id)
     tags_list = item.tags_string
-
     authors_list = []
     if item.type == 'book':
         authors_list = item.authors_string
-
     return render_template('item_description.html',
                            item=item,
                            tags_list=tags_list,
@@ -529,6 +559,7 @@ def item_description(item_id):
 
 
 @library.route('/add_copy/<int:item_id>', methods=['GET', 'POST'])
+@require_role('ADMIN')
 def add_copy(item_id):
     form = CopyAddForm()
     if form.validate_on_submit():
@@ -554,6 +585,7 @@ def add_copy(item_id):
 
 
 @library.route('/edit_copy/<int:copy_id>', methods=['GET', 'POST'])
+@require_role('ADMIN')
 def edit_copy(copy_id):
     copy = Copy.query.get_or_404(copy_id)
     item_id = copy.library_item_id
@@ -580,6 +612,7 @@ def edit_copy(copy_id):
 
 @library.route('/edit_profile/<int:user_id>',
                methods=['GET', 'POST'])
+@require_logged_in()
 def edit_profile(user_id):
     try:
         user = User.query.get(session['id'])
@@ -598,7 +631,6 @@ def edit_profile(user_id):
             return redirect(url_for('library.index'))
         except IntegrityError:
             abort(500)
-
     form.first_name.data = user.first_name
     form.surname.data = user.surname
     form.email.data = user.email
@@ -608,6 +640,7 @@ def edit_profile(user_id):
 
 
 @library.route('/reservations', methods=['GET', 'POST'])
+@require_role('ADMIN')
 def admin_dashboard():
     try:
         user = User.query.get(session['id'])
@@ -616,7 +649,6 @@ def admin_dashboard():
         abort(401)
     except Exception:
         abort(500)
-
     if request.method == 'GET':
         if not request.args or not request.args.get('search-query'):
             search_form = SearchForm(prefix="search")
@@ -629,7 +661,6 @@ def admin_dashboard():
                 reserv_page, 10, False)
             borrow_query = RentalLog.query.filter_by(book_status=2).order_by(
                 RentalLog._return_time.asc()).paginate(borrow_page, 10, False)
-
             return render_template('admin.html',
                                    reservations=reserv_query.items,
                                    borrows=borrow_query.items,
@@ -646,22 +677,18 @@ def admin_dashboard():
             borrow_form = BorrowForm(prefix="borrow")
             return_form = ReturnForm(prefix="return")
             query_str = request.args.get('search-query')
-
             reserv_page = request.args.get('page', 1, type=int)
             borrow_page = request.args.get('page', 1, type=int)
-
             reserv_query = RentalLog.query.filter_by(book_status=1).order_by(
                 RentalLog._reservation_begin.asc())
             reserv_filter = reserv_query.filter(
                 User.surname.ilike("%{}%".format(query_str))).paginate(
                 reserv_page, 10, False)
-
             borrow_query = RentalLog.query.filter_by(book_status=2).order_by(
                 RentalLog._return_time.asc())
             borrow_filter = borrow_query.filter(
                 User.surname.ilike("%{}%".format(query_str))).paginate(
                 borrow_page, 10, False)
-
             return render_template('admin.html',
                                    reservations=reserv_filter.items,
                                    borrows=borrow_filter.items,
@@ -671,28 +698,28 @@ def admin_dashboard():
                                    endpoint='library.admin_dashboard',
                                    search_form=search_form,
                                    borrow_form=borrow_form,
-                                   return_form=return_form,)
+                                   return_form=return_form)
 
     elif request.method == 'POST':
         search_form = SearchForm(prefix="search")
         borrow_form = BorrowForm(prefix="borrow")
         return_form = ReturnForm(prefix="return")
         if borrow_form.submit.data and borrow_form.validate_on_submit():
-
             copy_asset = request.args.get('asset')
-            borrow_item = Copy.query.filter_by(asset_code=copy_asset).first()
+            borrow_item = Copy.query.filter_by(asset_code=copy_asset). \
+                first_or_404()
             rental_log_change = RentalLog.query.filter_by(
-                copy_id=borrow_item.id).first()
+                copy_id=borrow_item.id).first_or_404()
             try:
+                borrow_item.available_status = BookStatus.BORROWED
                 rental_log_change.book_status = BookStatus.BORROWED
                 rental_log_change._borrow_time = datetime.now(tz=pytz.utc)
-                rental_log_change._return_time = (
-                    datetime.now(tz=pytz.utc) + timedelta(days=14))
+                rental_log_change._return_time = \
+                    (datetime.now(tz=pytz.utc) + timedelta(days=14))
                 db.session.commit()
-
             except exc.SQLAlchemyError:
                 abort(500)
-            flash('Item borrowed', 'Success!')
+            flash('Item borrowed')
             return redirect(url_for('library.admin_dashboard'))
 
         if return_form.submit.data and return_form.validate_on_submit():
@@ -700,24 +727,23 @@ def admin_dashboard():
             borrow_item = Copy.query.filter_by(asset_code=copy_asset).first()
             rental_log_change = RentalLog.query.filter_by(
                 copy_id=borrow_item.id).first()
-
             try:
+                borrow_item.available_status = BookStatus.RETURNED
                 rental_log_change.book_status = BookStatus.RETURNED
                 rental_log_change._borrow_time = None
                 rental_log_change._return_time = datetime.now(tz=pytz.utc)
                 db.session.commit()
-
             except exc.SQLAlchemyError:
                 abort(500)
-            flash('Item returned!', 'Success!')
+            flash('Item returned!')
             return redirect(url_for('library.admin_dashboard'))
-
     else:
         abort(500)
 
 
 @library.route('/edit_password/<int:user_id>',
                methods=['GET', 'POST'])
+@require_logged_in()
 def edit_password(user_id):
     try:
         user = User.query.get(session['id'])